perfetto: fix register/key truncation and unbounded recursion in protovm parser
- The map key read from a register was truncated to uint32 before the
  lookup, so two state entries whose 64-bit keys agree in the low 32
  bits silently collided.
- Register ids were truncated to uint8_t at the call sites, so e.g.
  dst_register=256 silently aliased register 0 instead of failing.
  Executor::{Read,Write}Register now take the program's uint32 wire
  value and reject out-of-range ids.
- ParseInstruction/ParseSelectRec recurse on producer-supplied data
  (the program's nesting and path length), with no depth bound: a
  deeply nested program could overflow the stack. Bound the recursion
  depth (128, generously above real programs).

+++ src/protovm/executor.h
@@ -62,8 +62,11 @@ class Executor {
   virtual StatusOr<RwProto::Cursor::RepeatedFieldIterator> IterateRepeatedField(
       RwProto::Cursor* dst,
       uint32_t field_id) const;
-  virtual StatusOr<uint64_t> ReadRegister(uint8_t reg_id) const;
-  virtual StatusOr<void> WriteRegister(const Cursors& cursors, uint8_t reg_id);
+  // reg_id is deliberately uint32_t (the program's wire type): narrowing it
+  // at the call site would silently alias registers (e.g. 256 -> 0). Out of
+  // range ids are rejected here.
+  virtual StatusOr<uint64_t> ReadRegister(uint32_t reg_id) const;
+  virtual StatusOr<void> WriteRegister(const Cursors& cursors, uint32_t reg_id);
   virtual StatusOr<void> Delete(RwProto::Cursor* dst) const;
   virtual StatusOr<void> Merge(Cursors* cursors, uint32_t flags) const;
   virtual StatusOr<void> Set(Cursors* cursors) const;
